Some cleanup/renaming/re-orging.
